refactor(models): expose IssueInfo/IssueState via vibe3.models public API
Add IssueInfo and IssueState to vibe3.models.__init__ and update all 53
import sites to use the public API instead of vibe3.models.orchestration
submodule imports. Remove model type re-exports from vibe3.orchestra to
enforce clear module boundary: models belong in vibe3.models, not in the
orchestra functional module.

- vibe3.models: add IssueInfo, IssueState to __init__ and __all__
- vibe3.orchestra: remove IssueInfo/IssueState from re-exports and __all__
- 51 files: replace submodule import with vibe3.models public API
- state_machine.py, label_service.py: split mixed imports (constants stay
  at vibe3.models.orchestration, model types move to vibe3.models)

Verified: 2515/2515 tests pass, mypy clean
