+++ src/vibe3/domain/handlers/supervisor_scan.py
@@ -15,7 +15,7 @@ from vibe3.clients.store_context import get_store
 from vibe3.config.orchestra_settings import load_orchestra_config
 from vibe3.domain.events.supervisor_apply import SupervisorIssueIdentified
 from vibe3.domain.handler_registry import register_handler
-from vibe3.models.orchestration import IssueInfo
+from vibe3.models import IssueInfo
 from vibe3.services.error_helpers import record_dispatch_failure_if_unexpected
 
 if TYPE_CHECKING:
