+++ src/vibe3/domain/protocols/flow_protocols.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING, Protocol
 
 if TYPE_CHECKING:
     from vibe3.clients.git_client import GitClient
-    from vibe3.models.orchestration import IssueInfo
+    from vibe3.models import IssueInfo
 
 
 class FlowManagerProtocol(Protocol):
