+++ src/vibe3/orchestra/protocols.py
@@ -11,7 +11,7 @@ The definition below is maintained for backward compatibility during migration.
 from typing import Protocol
 
 from vibe3.clients.git_client import GitClient
-from vibe3.models.orchestration import IssueInfo
+from vibe3.models import IssueInfo
 from vibe3.services.check_service import CheckResult
