+++ src/vibe3/services/blocked_state_io.py
@@ -14,8 +14,8 @@ from typing import TYPE_CHECKING, Literal
 from loguru import logger
 
 from vibe3.clients.github_client import GitHubClient
+from vibe3.models import IssueState
 from vibe3.models.issue_body import FlowStateProjection
-from vibe3.models.orchestration import IssueState
 from vibe3.services.blocked_state_types import BlockedState
 from vibe3.services.issue_body_service import merge_projection, parse_projection
 from vibe3.services.label_service import LabelService
