+++ src/vibe3/services/flow_orchestrator_service.py
@@ -25,7 +25,7 @@ from vibe3.services.task_service import TaskService
 
 if TYPE_CHECKING:
     from vibe3.config.orchestra_config import OrchestraConfig
-    from vibe3.models.orchestration import IssueInfo
+    from vibe3.models import IssueInfo
     from vibe3.services.orchestra_status_service import OrchestraSnapshot
