+++ src/vibe3/services/flow_rebuild_usecase.py
@@ -13,7 +13,7 @@ from vibe3.clients.git_client import GitClient
 from vibe3.clients.github_client import GitHubClient
 from vibe3.clients.sqlite_client import SQLiteClient
 from vibe3.config.orchestra_settings import load_orchestra_config
-from vibe3.models.orchestration import IssueInfo
+from vibe3.models import IssueInfo
 from vibe3.services.flow_cleanup_service import FlowCleanupService
 from vibe3.services.flow_orchestrator_service import FlowOrchestratorService
 from vibe3.services.flow_rebuild_postconditions import assert_rebuild_postconditions
@@ -132,8 +132,8 @@ class FlowRebuildUsecase:
         Post-rebuild: consistency is guaranteed (we just rebuilt), so we
         skip the classify step and go straight to clearing blocked state.
         """
+        from vibe3.models import IssueState
         from vibe3.models.flow import FlowState
-        from vibe3.models.orchestration import IssueState
         from vibe3.services.blocked_state_service import BlockedStateService
         from vibe3.services.flow_resume_resolver import infer_resume_label
