+++ src/vibe3/services/task_status_classifier.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 
 from enum import Enum
 
-from vibe3.models.orchestration import IssueState
+from vibe3.models import IssueState
 from vibe3.services.label_utils import has_manager_assignee
