+++ src/vibe3/services/task_status_service.py
@@ -3,9 +3,9 @@
 from dataclasses import dataclass
 from typing import Any, cast
 
+from vibe3.models import IssueState
 from vibe3.models.flow import FlowStatusResponse
 from vibe3.models.orchestra_config import OrchestraConfig
-from vibe3.models.orchestration import IssueState
 from vibe3.services.flow_service import FlowService
 from vibe3.services.orchestra_status_service import OrchestraSnapshot
 from vibe3.services.status_query_service import StatusQueryService, is_auto_task_branch
